The specification html page load time have again gotten too long. It is a travesty that in 21st century with our buncha-core gazillion Hz CPUs it is in order of tens of seconds to load a page. So I once again look if there is some low hanging fruit to help the situation.

This hides the page until everything is ready, to prevent the browser from trying to do incremental display of partially loaded page. At least for me the browser is largely unresponsive during the load time anyway, so it is no great loss.

Informal measurements:

Browser | Before [s] | After [s]
---------|------------ | -------------
Firefox  | 29              | 13
Chrome| 35              | 20

I also welcome some advice for aditional optimization from someone unlike me that knows what he's doing.

Some notes:
- Math translation takes ~ 2.5 s &mdash; that can be eliminated by doing it offline
- The style and layout is calulated twice &mdash; I have no idea why at the moment. It wastes ~8 s.
- I have a suspicion the TOC sidebar gives browsers some trouple. There may also be other tricky elements to layout (tables inside tables?)
- The binary images are inline in the html file. That is bound to create some performance problems...