@oddhack 2018-04-16
Why should having the images imbedded cause problems?