@krOoze 2018-04-16
@oddhack Because they need to be processed before the browser can continue constructing the HTML DOM and calculating layout. Theoretically the browser only needs to know the image dimensions to start calculating layout. So having whole image upfront may theoretically delay the start of the lengthy layout calculation.