@krOoze 2018-04-23
Updated with better handling of hidden style. Restricted only to `content` (shows header + toc immediately). Style the loading message.