@krOoze 2018-04-24
Fixed IE compatibility.