@krOoze 2018-04-24
I set up [krOoze/Vulkan-Docs/optimize_AOTA](https://github.com/krOoze/Vulkan-Docs/tree/optimize_AOTA) branch with all of the PRs.

And I set up unofficial [hosting](http://vulkan.hys.cz/1.1-extensions/vkspec.html) (which I believe is allowed under the conditional copyright). Test your favorite browser. Do **not** link to it &mdash; it is only temporary.