@kainino0x 2018-06-28
@oddhack Any chance landing some of these changes could get prioritized? I'm willing to help if there are any further changes needed to the PRs. There's also a chance I could ask some knowledgeable Chrome folks about optimizing this page further.

I also have a small, non-intrusive change that significantly improves the katex loading time, here:
https://github.com/KhronosGroup/Vulkan-Docs/compare/master...kainino0x:katex-faster-auto-render
Which I'll open a PR for after I poke at it some more.