@krOoze 2018-06-28
> There's also a chance I could ask some knowledgeable Chrome folks about optimizing this page further.

OMG, yes, please, please. With the above PRs it is almost there. Though in my naivette I don't see a fundamental reason why it should take >1 s to load.

> I also have a small, non-intrusive change that significantly improves the katex loading time, here:
master...kainino0x:katex-faster-auto-render

Should probably get rid of the added `<span>` when replaced.
Block equations should be `<div>`, not `<span>`?

Anyway, PR #704 would get rid of all of that by translating it offline\server-side. Should also avoid triggering style and layout recalculation.