@kainino0x 2018-06-28
> In my naivette I don't see a fundamental reason why it should take >1 s to load though.

I think it is inevitable at the sheer size of the generated DOM, except, probably, by hiding sections by default. However, it's possible that the complexity of layout could be reduced enough to make it quicker.

> Should probably get rid of the added <span> when replaced.

This might affect the surrounding DOM element, which could make it a lot more expensive. However, I will look into it.

> Block equations should be <div>, not <span>?

They're separately surrounded by a div (`<div><span></span></div>`) I believe, so this isn't necessary.

> Anyway, PR #704 would get rid of all of that by translating it offline\server-side. Should also avoid triggering style and layout recalculation.

I agree that it is probably a better solution - I intend to take a closer look at that patch. However I'd love to get a "quick fix" in sooner (doesn't require Node, for one thing).