@krOoze 2018-06-28
> I think it is inevitable at the sheer size of the generated DOM

I would understand it on a cellphone or tablet. But on desktop processing measly 8 MB of data almost seems like one of the easiest types of tasks... Not to mention that HTML size is quite inflated by being human-readableish.