@kainino0x 2018-06-28
on my (fast) workstation:
* With (my) script, with style: ~6.4s
* No script, with style: ~3.2s
* No script, no style (except for the inline style attributes): ~2.5s

So improving the styles could saves some time, but the real killer is the script itself and relayout caused by the script.

I do think it might be preferable to un-inline the images, if we could. I read something yesterday saying that data: uri images require re-decoding on every load, which I think could particularly impact closing and reopening vkspec in a single browser session (something I do a lot). It might reduce parsing time a little; not sure.