@kainino0x 2018-06-28
> > Should probably get rid of the added `<span>` when replaced.

This turns out to be cheap, and seems to save about 15-20% of relayout time! Thanks for the tip!