@kainino0x 2018-06-28
no script, img src="data:...": ~3.1s
no script, img src="": ~2.6s

This could be because the `<img>` tags in vkspec do not have width and height. This is bad because it causes layout to change when the image is loaded and the browser suddenly knows the width and height of the image.

Fixing this would require some automation, so not sure how to do it.