@krOoze 2018-06-28
@kainino0x The `asciidoctor` should be doing it itself if you remove the `:data-uri:` at https://github.com/KhronosGroup/Vulkan-Docs/blob/master/vkspec.txt#L7