@kainino0x 2018-06-28
I actually meant that adding the width= and height= attributes would require automation - I don't know if asciidoctor can do that, plus it's a bit tricky to make it work in a way that is "responsive" to the window size.

But thanks for the pointer! I checked the performance of de-inlining the images, and I also found [`[%inline]`](https://github.com/asciidoctor/asciidoctor-bespoke#inserting-svgs).

no script, img src="data:...": 3.1s
no script, img src="images/...": 2.9s
no script, `[%inline]`: 3.05?s

So there's a tiny improvement here. This is on `localhost`, but with Chrome's "Fast 3G" emulation:

3g, no script, img src="data:...": 30.1s
3g, no script, img src="images/...": 28.2s
3g, no script, `[%inline]`: 28.0s

This is probably because the plaintext svg images are smaller than the base64 encoded images, so that might be the whole savings.
(I don't have gzip on my local server, which would also change the results.)

It seems from this that the `[%inline]` method would be the way to go.

[Inline SVG](https://caniuse.com/#feat=svg-html5) is about as well supported as [SVG itself](https://caniuse.com/#feat=svg).

---

I also tried [`svgo`](https://www.npmjs.com/package/svgo), an SVG optimizer, to minify all of the SVGs:

no script, img src="data:...": vkspec.html=5445607 B (697726 B gz)
no script, `[%inline]`: vkspec.html=5041131 B (609636 B gz)
no script, `[%inline]`, svgo: vkspec.html=4321077 B (546450 B gz)

And [`html-minifier`](http://kangax.github.io/html-minifier/) to minify the inline HTML/CSS/JS:

no script, `[%inline]`, svgo + html-minifier: vkspec.html=4000019 B (509007 B gz)

The minifiers seem to give a tiny load time improvement even without 3G emulation.