@kainino0x 2018-06-29
Opened https://github.com/KhronosGroup/Vulkan-Docs/pull/734 for `[%inline]`.