@kainino0x 2018-06-29
I think my preference would be to land #704 but not #702. Reason being:

#704 (at least for me) significantly reduces the page load time. However the most important thing it does is avoid ever forcing a relayout - layout happens incrementally while the page loads, and the page never freezes (stops being scrollable).

#702 (at least for me, in Chrome 67) causes a huge relayout AFTER the page has loaded - because the entire page still has to be laid out (it couldn't be done incrementally). Plus, it makes the page unusable (and unsearchable) while loading, so I think we shouldn't use it.

#708 also prevents some big relayouts (freezes) while the page is loading, especially on a slower connection where it takes longer to get the fonts in the first place. It doesn't improve loading time a lot for me, but it does greatly improve responsiveness.

@krOoze, what do you think?